backend/requirements.txt
# ─── AuraGrade Backend Dependencies (Pinned for Reproducibility) ─────
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
google-genai>=1.0.0,<2.0.0
python-multipart>=0.0.9
supabase>=2.0.0,<3.0.0
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
pymupdf>=1.24.0,<2.0.0
opencv-python-headless>=4.9.0,<5.0.0
numpy>=1.26.0,<3.0.0
pinecone>=5.0.0,<6.0.0
pinecone-text>=0.9.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
mcp>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=8.2.0,<10.0.0
PyJWT>=2.8.0,<3.0.0
slowapi>=0.1.9,<1.0.0
